// java-components/management-console/src/main/webui/src/services/openapi/models/RepositoryInfo.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { TagMapping } from './TagMapping';
export type RepositoryInfo = {
    type?: string;
    uri?: string;
    path?: string;
    private?: boolean;
    tagMapping?: Array<TagMapping>;
    uriWithoutFragment?: string;
};